rename phaser folder
