+++ src/phaser/card/CardData.ts
@@ -1,0 +1,7 @@
+import { Suit } from "phaser/constants/deck";
+
+export type CardData = {
+  id: string;
+  suit: Suit;
+  value: number;
+};
