+++ src/phaser/card/state/CardState.ts
@@ -1,0 +1,7 @@
+import { PileId } from "phaser/constants/table";
+
+export type CardState = {
+  pile: PileId;
+  position: number;
+  flipped: boolean;
+};
