+++ src/phaser/constants/table.ts
