+++ src/phaser/pile/PileData.ts
@@ -1,0 +1,5 @@
+import { PileId } from "phaser/constants/table";
+
+export type PileData = {
+  id: PileId;
+};
