--- src/phasor/card/CardData.ts
@@ -1,7 +1,0 @@
-import { Suit } from "@phasor/constants/deck";
-
-export type CardData = {
-  id: string;
-  suit: Suit;
-  value: number;
-};
