--- src/phasor/card/state/CardState.ts
@@ -1,7 +1,0 @@
-import { PileId } from "@phasor/constants/table";
-
-export type CardState = {
-  pile: PileId;
-  position: number;
-  flipped: boolean;
-};
