--- src/phasor/pile/PileController.ts
@@ -1,17 +1,0 @@
-import { PileView } from "@phasor/pile/PileView";
-import { Pile } from "@phasor/pile/state/Pile";
-
-export class PileController {
-  constructor(
-    public readonly model: Pile,
-    public readonly view: PileView,
-  ) {}
-
-  setTint(tint: number): void {
-    this.view.setTint(tint);
-  }
-
-  clearTint(): void {
-    this.view.clearTint();
-  }
-}
