--- src/phasor/pile/PileData.ts
@@ -1,5 +1,0 @@
-import { PileId } from "@phasor/constants/table";
-
-export type PileData = {
-  id: PileId;
-};
