--- src/phasor/pile/state/Pile.ts
@@ -1,14 +1,0 @@
-import { PileId } from "@phasor/constants/table";
-import { PileData } from "@phasor/pile/PileData";
-
-export type Pile = {
-  data: PileData;
-};
-
-export function createPile(id: PileId): Pile {
-  return {
-    data: {
-      id: id,
-    },
-  };
-}
